More sane query for modifying addons

+++ src/GQL/Addons/Controller.ts
@@ -1,44 +1,11 @@
 import { GraphQLError } from "graphql";
-import { PersonRole, type PropertyAddonType } from "@prisma/client";
+import { PersonRole } from "@prisma/client";
 import { Prisma } from "DB/Client";
 import { Permission } from "Tools/Permission";
 import type { Session } from "Types/GraphQL";
+import type { IModifyAddons } from "./Types";
 
 export class AddonController {
-  public static createAddon = async (
-    propertyId: number,
-    type: PropertyAddonType,
-  ) => {
-    await Prisma.transact(async client => {
-      try {
-        await client.propertyAddon.create({
-          data: {
-            type,
-            propertyId,
-          },
-        });
-      } catch (error) {
-        throw new GraphQLError(
-          "This addon already exists. Please refresh your page",
-        );
-      }
-    });
-    return this.getAddons(propertyId);
-  };
-
-  public static createAddons = async (
-    propertyId: number,
-    types: PropertyAddonType[],
-  ) => {
-    await Prisma.transact(async client => {
-      await client.propertyAddon.createMany({
-        data: types.map(type => ({ propertyId, type })),
-        skipDuplicates: true,
-      });
-    });
-    return this.getAddons(propertyId);
-  };
-
   public static getAddons(propertyId: number) {
     return Prisma.transact(client => {
       return client.propertyAddon.findMany({
@@ -53,34 +20,35 @@ export class AddonController {
     });
   }
 
-  public static deleteAddon = async (id: number, propertyId: number) => {
+  public static modifyAddons = async <
+    T extends Omit<IModifyAddons, "organizationId">,
+  >({
+    propertyId,
+    deletions,
+    additions,
+  }: T) => {
     await Prisma.transact(async client => {
       try {
-        await client.propertyAddon.delete({
-          where: { id },
-        });
-      } catch (error) {
-        throw new GraphQLError(
-          "Something went wrong when deleting this addon. Please refresh your page.",
-        );
-      }
-    });
-    return this.getAddons(propertyId);
-  };
-
-  public static deleteAddons = async (ids: number[], propertyId: number) => {
-    await Prisma.transact(async client => {
-      try {
-        await client.propertyAddon.deleteMany({
-          where: {
-            id: {
-              in: ids,
+        if (deletions.length) {
+          await client.propertyAddon.deleteMany({
+            where: {
+              id: {
+                in: deletions,
+              },
             },
-          },
-        });
+          });
+        }
+        if (additions.length) {
+          await client.propertyAddon.createMany({
+            data: additions.map(addon => ({
+              propertyId,
+              type: addon,
+            })),
+          });
+        }
       } catch (error) {
         throw new GraphQLError(
-          "Something went wrong when deleting this addon. Please refresh your page.",
+          "Something went wrong when updating your addons. Would you mind refreshing the page and trying again?",
         );
       }
     });
@@ -101,7 +69,7 @@ export class AddonController {
         ))
       ) {
         throw new GraphQLError(
-          "You do not have permission to modify the configuration of this property",
+          "You do not have permission to modify the addons of this property",
         );
       }
       return callback(...args);
